Update output thread
Simplify output packet format, so that it just includes a header of
mcnt, packet offset, payload length and xengine id.
Possibly want to add an explicit channel number (although this
is mostly redundant with xeng_id).

Capturing output data with antenna-varying and freq-varying
test vectors from SNAPs all look OK.

TODO: scriptify formal reproducible tests, including software-based
noise injection.

+++ src/paper_databuf.h
@@ -242,7 +242,7 @@ typedef uint8_t paper_output_header_cache_alignment[
 typedef struct paper_output_block {
   paper_output_header_t header;
   paper_output_header_cache_alignment padding; // Maintain cache alignment
-  float data[N_OUTPUT_MATRIX];
+  int32_t data[N_OUTPUT_MATRIX];
 } paper_output_block_t;
 
 typedef struct paper_output_databuf {
